Update patch set 1
Patch Set 1: -Code-Review -Verified

Well - its a basic question for what this code is actually there.
So IMHO this is only useful for devs and even then a
"small" vsync miss would not be visible at all. Actually adding this log would delay it even more :)
If something goes terrible wrong with vsync you will get anyway delays that are much larger. The conversion to ms before comparing will just "round up" and it will not hit it for delays up to 1ms which happens surprisingly often at least on find5. I would actually opt to make this dependent to VSYNC_DEBUG or even add a property to enable it. But this is not really important and for now it makes sense to have it there to see issues with new devices earlier.

Patch-set: 1
Reviewer: Gerrit User 1000026 <1000026@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0