Update patch set 1
Patch Set 1:

There is a slight possibility this may be the only option for maintainers that have been fully kondicked - it seems like jflte has lots of weird problems if you just pull the kernel side of this, I'm waiting for further confirmation from Jake.

It may be easier to have this in here than to make people completely rework their GPU drivers for non-frankendisplayed blobs.  I'm looking for more info to confirm that though, I'd prefer to keep the frankendisplay mess out if there's no good justification for it.

Patch-set: 1