Update patch set 1
Patch Set 1: Code-Review+1

these are all the commits prior to the jflte adreno/gpu update that broke everything. these commits as they are here fix everything related to that update on jflte.

I have these in an UNsquashed state locally, and will push them up individually if you guys want me to, just figured this would be easier for everyone to test squashed up.

want to get this in quick since this was the last major issue with jflte! :)

Pretty sure I got all the maintainers who are using display-caf, if I forgot please add. I'll send this link out on the mailing list as well (which was a great idea Andrew)

Patch-set: 1
Label: Code-Review=+1