Update patch set 1
Patch Set 1:

The patch is done by me, without it the screen starts to tear a lot on S4 Mini, which is msm8930 but currently uses code from msm8960. Without the patch the logcat is flooded with messages like this ones:

E/qdhwcomposer(  299): hwc_sync: ioctl MSMFB_BUFFER_SYNC failed, err=Invalid argument
E/qdhwcomposer(  299): hwc_sync: acq_fen_fd_cnt=3 flags=0 fd=20 dpy=0 numHwLayers=4

I follow that log piece till this file I modified, where new changes were introduced from this commit:

https://github.com/omnirom/android_hardware_qcom_display-caf/commit/af38de8b0ec0fc1f41d1a7ead7981b5462261b21#diff-a173c249cb0ea0c9f46a57124f0f0158R1066

The modifications from lines 1066 to 1068 on hwc_utils.cpp make that issue to appear for me. Notice that if you follow the logic table of that checks, the only way to pass it is that list->hwLayers[i].compositionType must to be equal to HWC_OVERLAY and HWC_BLIT at the same, something I that cannot be possible, because that var is set depending of MDP version here:
https://github.com/omnirom/android_hardware_qcom_display-caf/blob/android-4.4/libhwcomposer/hwc_copybit.cpp#L203

So I modified the and(&&)/or(||) in order to allow passing the check in both cases.
Notice that the commit I linked and that you merged haven't been merged on CM display-caf tree, where I still can reproduce a build without the issue and pass the check:
https://github.com/CyanogenMod/android_hardware_qcom_display-caf/blob/cm-11.0/libhwcomposer/hwc_utils.cpp#L1170

Patch-set: 1