Update patch set 1
Patch Set 1:

I've been checking a bit more the Cyano tree, and they merged the same commit you merged and that broke it for me, but the partially reverted to the older state on this commit:
https://github.com/CyanogenMod/android_hardware_qcom_display-caf/commit/da4a0f19eb88332e3be3b20146a8e0df6e2a40ff

That confirms that my commit won't broke it, because anyway copybit waiting for fences is not required, but probably you would prefer to merge that other. Will test that commit and report later if it works for me.

Patch-set: 1