Update patch set 1
Patch Set 1:

Yea, I mean that's really the only thing I've come to realize this is. Like I said, everything works perfectly fine for jflte as it sits right now with it's repositories on github. This (afaik) would only allow the flo gpu blobs to be used, which in turn would allow users to have the option to flash custom kernels. I know that is completely unsupported but that's the last major gripe and complaint.I've been seeing for jflte.

It really doesn't matter to me either way because 1) it works without this stuff  (No flo) and 2) *fuck* CM for setting the standard for all jflte kernel development to depend on this stuff.

Patch-set: 1